# samples/sample_openapi.yaml
openapi: 3.0.3
info:
  title: Sample User API
  version: 0.4.0
  description: Sample OpenAPI spec for parser, generator, executor, and local demo API development.

servers:
  - url: http://localhost:8000

paths:
  /users/{userId}:
    parameters:
      - name: userId
        in: path
        required: true
        description: Unique identifier of a user.
        schema:
          type: integer
          minimum: 1
          example: 10
    get:
      summary: Get user by id
      operationId: getUserById
      parameters:
        - name: includePosts
          in: query
          required: false
          description: Include recent posts in the response.
          schema:
            type: boolean
            example: true
        - name: X-Trace-Id
          in: header
          required: false
          description: Trace identifier for debugging.
          schema:
            type: string
            example: trace-123
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "422":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /users:
    post:
      summary: Create a user
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "422":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
          minLength: 3
        email:
          type: string
          format: email
        age:
          type: integer
          minimum: 18
        role:
          type: string
          enum: [admin, member]
      required:
        - id
        - username
        - email
        - age
        - role

    CreateUserRequest:
      type: object
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 20
          example: alice
        email:
          type: string
          format: email
          example: alice@example.com
        age:
          type: integer
          minimum: 18
          maximum: 120
          example: 25
        role:
          type: string
          enum: [admin, member]
          example: member
      required:
        - username
        - email
        - age
        - role

    ErrorResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message